Import a function from bit64 so the namespace is loaded
Fixes test errors due to failure to dispatch on integer64 objects.

This took a long time to figure out, the tests were passing
interactively because `devtools::load_all()` calls
`pkgload::check_dep_version()` on all package dependencies, which
implicitly loads the packages. However testhat::test_check() does not do
this implicit loading of bit64, so the dispatch was not happening.