Fix translation of missing values for integer types
Previously we were using the null_sentinel interface when calling
nanodbc::statement::bind(). However that doesn't seem to be working
properly, perhaps it is picking up the wrong method when used with 32
bit integers?

Explicitly using the null vector interface like that used for doubles
works however, so we will switch to using that instead.

Fixes #119