Add odbcConnectionColumns generic
This is an exported function that is envisioned as a thin wrapper around SQLColumns.
Default method points to renamed/reordered connection_sql_columns.

We hope to call this method in dbWriteTable, to set parameter descriptions prior to binding values.  Given that drivers may have idiosyncratic implementations of SQLColumns (for example, some may not offer the ability to call this function on a table outside of the current catalog) offer an S4 method giving the end-user to write an implementation that works for them.