Feature/describe prior to write (#313)
* nanodbc: Add describe_parameters function as alternative to SQLDescribeParam

Backport of e99606bc8d01257c7c1316dd06f3a9c30e0a71fd

* Add odbcConnectionColumns generic

This is an exported function that is envisioned as a thin wrapper around SQLColumns.
Default method points to renamed/reordered connection_sql_columns.

We hope to call this method in dbWriteTable, to set parameter descriptions prior to binding values.  Given that drivers may have idiosyncratic implementations of SQLColumns (for example, some may not offer the ability to call this function on a table outside of the current catalog) offer an S4 method giving the end-user to write an implementation that works for them.

* Add result_describe_parameters R/cpp function

Calls nanodbc::describe_parameters

* dbWriteTable: Attempt to describe parameters prior to binding/inserting

Call to result_describe_parameters, which in turn calls nanodbc::describe_parameters, is wrapped
in a tryCatch block - if there is a failure, we fallback to the nanodbc code-path where binding uses a call to SQLDescribeParam

* odbcConnectionColumns: Respect column_names parameter

Also fixup documentation

* dbWriteTable: Use concise data types when describing parameters

* code-review: Incorporate feedback

- Wrap conditional expressions in braces
- Avoid using numeric indexing when re/ordering output of
  connection_sql_columns
- Bugfix: Appropriate index passed down to
  nanodbc::statement::describe_parameters

* tests: Extend test_roundtrip

Check the case when, via `dbWriteTable`, we are writing a data.frame with columns
ordered differently than the table we are writing to.  In `dbWriteTable` we
attempt to describe parameters (types, length) of the INSERT query using
information about the columns of the table being written to.  Some care is needed
to make sure that the table column descriptions get paired with the appropriate parameters.

* code-review ctd: No re-naming/-ordering in odbcConnectionColumns

Also move test checking for out-of-order column write to tests-PostgreSQL

* Add note to NEWS