Fix wide character issues with the macOS binary
My hypothesis is that the macOS builder has iODBC in the system include
directory. Because this directory comes in a `-I` directive _before_ the
one added by `PKG_CXXFLAGS` we were getting the iODBC `sql.h` header
instead of the unixodbc one. iODBC uses a different width of SQLWCHAR,
so this is why we were getting messed up results. Putting the proper
include directory in `PKG_CPPFLAGS` instead of `PKG_CXXFLAGS` makes it
come _before_ the system includes, so we should pick up the unixodbc one
we install by homebrew, with the proper width.

Fixes #283