dbGetQuery: on.exit ensures result is destroyed
The result is local to the method, and as is, it will get released
(and associated resources - nanodbc::statement,for example - destroyed)
on.exit with a call to `dbClearResult` and failing that once the
garbage collector engages.

If the result fetch fails/throws:
  * dbClearResult may not release the result, as the exception handler will
    mark the `current` result associated with the connection to null.
  * We may need to destroy associated resources in a more deterministic
    fashion than leaving it to the garbage collector, as the nandobc
    destructors themselves may free remote resources / reset processing
    associated with the failed statement.

This patch ensures we do that by calling `result_release` in lieu of
`dbClearResult` on exit.