backport: nanodbc: Allow callers to pass pre-connection attributes
This is a backport of nanodbc/commit/b0fa192eb7289bba5450f90ad57075dec3a6581c

The solution there is restricted to C++17 ( using variant ).  Here we use a slightly
simpler definition of connection::attribute that is C++11 friendly.  In particular,
where the upstream solution uses a std::variant for the connection attribute payload,
here we the caller passes a pointer to a resource.