fix dbExistsTable and sqlCreateTable for Teradata
Temporary (volatile) tables in Teradata are not found by `connection_sql_tables`. The edge case was not handled by `dbExistsTable`. A specific SQL query needs to be run.

sqlCreateTable needs additional arguments to be useful (index, partition_by). In my company, most tables are specified as "multiset". It is added as the default.