Use SQLINTEGER in get_info()
Instead of `unsigned short`, which could be allocated
or aligned differently by different compilers.

There are other places in nanodbc that use `unsigned short`
in the protocol, so potentially those are broken as well.

I am not very familiar with the ODBC toolchain, so it
is possible that this is not the best fix, but nevertheless
this is the step that causes the memory error and the
crash, so this needs to change.

For #624.