Use SQLUINTEGER in get_info() (#677)
* Use SQLINTEGER in get_info()

Instead of `unsigned short`, which could be allocated
or aligned differently by different compilers.

There are other places in nanodbc that use `unsigned short`
in the protocol, so potentially those are broken as well.

I am not very familiar with the ODBC toolchain, so it
is possible that this is not the best fix, but nevertheless
this is the step that causes the memory error and the
crash, so this needs to change.

For #624.

* get_info(): use SQLUINTEGER

That's what we need, not SQLINTEGER. See SQL_GETDATA_EXTENSIONS
at https://learn.microsoft.com/en-us/sql/odbc/reference/syntax/sqlgetinfo-function?view=sql-server-ver16.

* Add get_info() template instance for SQLUINTEGER

Needed on Windows, where it is `unsigned long`, and
that is different from `uint32_t` and `uint64_t`,
apparently.

* get_info(): only instantiate SQL types

Adding an instance to SQLUINTEGER fails on
Linux. Let's try to only add instances for
the SQL types now.