Read Workbench-managed Databricks credentials from databricks.cfg.
Previously odbc::databricks() used .rs.api.getDatabricksToken() to check
for Workbench-managed credentials, which only works inside the RStudio
console. With this commit we look at the generated config file instead,
as we do for Snowflake in odbc::snowflake(). This gives background jobs
and, importantly, the Quarto and RMarkdown rendering panes access to
these credentials as well.

Unit tests are included.

Closes #803.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>