Add support for using Connect-managed Databricks service principals.
Important: this commit implements a client for a *draft* API of an
*unreleased* Connect feature. It is intended to serve as an internal
reference for what client implementations could look like. It cannot and
should not be merged before this API is released (if it ever is).

---

This commit allows retrieving OAuth credentials for a Databricks service
principal managed by Posit Connect. Like viewer-based credentials, this
involves an OAuth exchange flow with the Connect server, but unlike
viewer credentials there is no user session involved; this is a pure
service account model.

Unit tests and documentation are TODO.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>